+++ example/src/renderer/examples/basic/VideoDecoder/VideoDecoder.tsx
@@ -5,6 +5,8 @@ import {
   IRtcEngineEventHandler,
   IRtcEngineEx,
   LogFilterType,
+  RtcConnection,
+  RtcStats,
   VideoSourceType,
   createAgoraRtcEngine,
 } from 'agora-electron-sdk';
@@ -14,13 +16,15 @@ import {
   BaseAudioComponentState,
   BaseComponent,
 } from '../../../components/BaseComponent';
-import { AgoraTextInput } from '../../../components/ui';
+import { AgoraButton, AgoraTextInput } from '../../../components/ui';
 import Config from '../../../config/agora.config';
 import { askMediaAccess } from '../../../utils/permissions';
 
 interface State extends BaseAudioComponentState {
+  token2: string;
+  uid2: number;
   fps: number;
-  decodeRemoteUserUid: number;
+  joinChannelExSuccess: boolean;
 }
 
 export default class VideoDecoder
@@ -38,9 +42,10 @@ export default class VideoDecoder
       token: Config.token,
       uid: Config.uid,
       joinChannelSuccess: false,
-      decodeRemoteUserUid: 7,
+      token2: '',
+      uid2: 0,
       remoteUsers: [],
-      startPreview: false,
+      joinChannelExSuccess: false,
     };
   }
 
@@ -73,9 +78,6 @@ export default class VideoDecoder
     // Need to enable video on this case
     // If you only call `enableAudio`, only relay the audio stream to the target channel
     this.engine.enableVideo();
-    // Start preview before joinChannel
-    this.engine?.startPreview();
-    this.setState({ startPreview: true });
   }
 
   /**
@@ -105,6 +107,41 @@ export default class VideoDecoder
   }
 
   /**
+   * Step 2-1(optional): joinChannelEx
+   */
+  protected joinChannelEx = () => {
+    const { channelId, token2, uid2 } = this.state;
+    if (!channelId) {
+      this.error('channelId is invalid');
+      return;
+    }
+    if (uid2 <= 0) {
+      this.error('uid2 is invalid');
+      return;
+    }
+    // publish screen share stream
+    this.engine?.joinChannelEx(
+      token2,
+      { channelId, localUid: uid2 },
+      {
+        autoSubscribeAudio: true,
+        autoSubscribeVideo: true,
+        publishMicrophoneTrack: true,
+        publishCameraTrack: true,
+        clientRoleType: ClientRoleType.ClientRoleBroadcaster,
+      }
+    );
+  };
+
+  /**
+   * Step 2-2(optional): leaveChannelEx
+   */
+  leaveChannelEx = () => {
+    const { channelId, uid2 } = this.state;
+    this.engine?.leaveChannelEx({ channelId, localUid: uid2 });
+  };
+
+  /**
    * Step 4: leaveChannel
    */
   protected leaveChannel() {
@@ -119,24 +156,55 @@ export default class VideoDecoder
     this.engine?.release();
   }
 
+  onJoinChannelSuccess(connection: RtcConnection, elapsed: number) {
+    const { uid2 } = this.state;
+    if (connection.localUid === uid2) {
+      this.setState({ joinChannelExSuccess: true });
+    } else {
+      this.setState({ joinChannelSuccess: true });
+    }
+    this.info(
+      'onJoinChannelSuccess',
+      'connection',
+      connection,
+      'elapsed',
+      elapsed
+    );
+  }
+
+  onLeaveChannel(connection: RtcConnection, stats: RtcStats) {
+    const { uid2 } = this.state;
+    if (connection.localUid === uid2) {
+      this.setState({ joinChannelExSuccess: false });
+    } else {
+      this.setState({ joinChannelSuccess: false });
+    }
+    this.info('onLeaveChannel', 'connection', connection, 'stats', stats);
+    this.setState(this.createState());
+  }
+
   protected renderUsers(): ReactElement | undefined {
-    let { decodeRemoteUserUid, startPreview, remoteUsers, joinChannelSuccess } =
-      this.state;
+    let { remoteUsers, joinChannelSuccess, joinChannelExSuccess } = this.state;
     return (
       <>
-        {!!startPreview || joinChannelSuccess
-          ? this.renderUser({
-              sourceType: VideoSourceType.VideoSourceCamera,
-            })
-          : undefined}
-        {!!startPreview || joinChannelSuccess
+        {joinChannelSuccess
           ? remoteUsers.map((item) =>
               this.renderUser({
                 uid: item,
                 // Use WebCodecs to decode video stream
-                // only support one remote stream to decode at the same time for now
-                useWebCodecsDecoder: item === decodeRemoteUserUid,
-                enableFps: item === decodeRemoteUserUid,
+                useWebCodecsDecoder: true,
+                enableFps: true,
+                sourceType: VideoSourceType.VideoSourceRemote,
+              })
+            )
+          : undefined}
+        {joinChannelExSuccess
+          ? remoteUsers.map((item) =>
+              this.renderUser({
+                uid: item,
+                // Use WebCodecs to decode video stream
+                useWebCodecsDecoder: true,
+                enableFps: true,
                 sourceType: VideoSourceType.VideoSourceRemote,
               })
             )
@@ -145,22 +213,50 @@ export default class VideoDecoder
     );
   }
 
-  protected renderConfiguration(): ReactElement | undefined {
+  protected renderChannel(): ReactElement | undefined {
+    const { channelId, joinChannelSuccess, joinChannelExSuccess } = this.state;
     return (
       <>
         <AgoraTextInput
-          disabled={this.state.joinChannelSuccess}
+          onChangeText={(text) => {
+            this.setState({ channelId: text });
+          }}
+          placeholder={`channelId`}
+          value={channelId}
+        />
+        <AgoraButton
+          title={`${joinChannelSuccess ? 'leave' : 'join'} Channel`}
+          disabled={joinChannelExSuccess}
+          onPress={() => {
+            joinChannelSuccess ? this.leaveChannel() : this.joinChannel();
+          }}
+        />
+      </>
+    );
+  }
+
+  protected renderConfiguration(): ReactElement | undefined {
+    let { joinChannelExSuccess, uid2, joinChannelSuccess } = this.state;
+    return (
+      <>
+        <AgoraTextInput
+          editable={!joinChannelExSuccess && !joinChannelSuccess}
           onChangeText={(text) => {
             if (isNaN(+text)) return;
             this.setState({
-              decodeRemoteUserUid:
-                text === '' ? this.createState().decodeRemoteUserUid : +text,
+              uid2: text === '' ? this.createState().uid2 : +text,
             });
           }}
           numberKeyboard={true}
-          placeholder={`useWebCodecsDecoder Uid (defaults: ${
-            this.createState().decodeRemoteUserUid
-          })`}
+          placeholder={`uid2 (must > 0)`}
+          value={uid2 > 0 ? uid2.toString() : ''}
+        />
+        <AgoraButton
+          title={`${joinChannelExSuccess ? 'leave' : 'join'} channelEx`}
+          disabled={joinChannelSuccess}
+          onPress={
+            joinChannelExSuccess ? this.leaveChannelEx : this.joinChannelEx
+          }
         />
       </>
     );